Struct assembly_pack::pk::reader::PackFile

pub struct PackFile<T> { /* private fields */ }

Available on crate feature pk only.

A low level pack file reader

Implementations

impl<T> PackFile<T>
where T: Seek + BufRead,

pub fn open(inner: T) -> Self

Open a file from a stream

pub fn into_inner(self) -> T

Return the inner reader

pub fn get_ref(&self) -> &T

Get a reference to the inner reader

pub fn get_mut(&mut self) -> &T

Get a mutable reference to the inner reader

pub fn check_magic(&mut self) -> Result<()>

Check for the magic bytes at the beginning of the file

pub fn get_header(&mut self) -> Result<PKTrailer>

Load the header from the end of the file

pub fn get_entry(&mut self, addr: u32) -> Result<PKEntry>

Load the header from the end of the file

pub fn get_entry_accessor(self, addr: u32) -> Result<PackEntryAccessor<T>>

Get an random access wrapper for the entries

pub fn get_entry_list(&mut self, addr: u32) -> Result<Vec<PKEntry>>

Get a list of all entries

pub fn get_file_stream<'b>( &'b mut self, entry: PKEntry, ) -> PackStreamReader<'b, T>

Get a boxed reader for the file stream

pub fn get_file_data( &mut self, entry: PKEntry, ) -> Result<PackDataStream<'_, T>, Error>

Get some object with a read trait representing the data